Move MethodMapping methods to its own class.
